requirements.txt
mkdocs-material==9.6.21
mkdocs-macros-plugin==1.4.0
pydantic==2.11.9
